// src/types/user.ts
export interface UserProfile {
  name: string;
  age: number;
  currentRole: string;
  skills: string[];
  dreamCareer: string;
  goals: string[];
}

export interface UserDraftProfile {
  name: string;
  age: string;
  currentRole: string;
  skills: string[];
  dreamCareer: string;
  goals: string[];
}